Implement assembler directives for RISC-V

Implement the core set of Pep/10 commands, minus import/export, input/output and SCALL.

Additionally, implement the following specifically for RISC-V
- HALF, WORD
- ASCIZ, STRING
- ALIGN, P2ALIGN, BALIGN
- TEXT, BSS, DATA, RODOTA
- GLOBAL, WEAK, LOCAL, HIDDEN

We do not yet support `.OPTION`, which will be added in the near future.

+++ core/core/compile/ir_linear/line_base.cpp
@@ -22,4 +22,4 @@ void pepp::tc::LinearIR::insert(std::unique_ptr<AAttribute> attr) {
   *link = std::move(node);
 }
 
-std::optional<u16> pepp::tc::LinearIR::object_size(u16) const { return std::nullopt; }
+std::optional<u64> pepp::tc::LinearIR::object_size(u64) const { return std::nullopt; }
